Giant update
Added archer and bomber kit, and tons more….

+++ src/me/theremixpvp/ckitpvp/PDUtils.java
@@ -3,20 +3,10 @@ package me.theremixpvp.ckitpvp;
 import java.io.File;
 import java.io.IOException;
 import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
 
-import org.bukkit.Bukkit;
-import org.bukkit.Material;
-import org.bukkit.OfflinePlayer;
 import org.bukkit.configuration.file.FileConfiguration;
 import org.bukkit.configuration.file.YamlConfiguration;
-import org.bukkit.enchantments.Enchantment;
 import org.bukkit.inventory.Inventory;
-import org.bukkit.inventory.ItemStack;
-import org.bukkit.inventory.meta.ItemMeta;
-import org.bukkit.material.MaterialData;
 import org.bukkit.plugin.Plugin;
 
 public class PDUtils {
@@ -74,7 +64,7 @@ public class PDUtils {
 		File f = new File(p.getDataFolder() + "/playerdata/" + name + ".yml");
 		if(f.exists()) {
 		FileConfiguration cfg = YamlConfiguration.loadConfiguration(f);
-		double credits = cfg.getDouble("credits");
+		int credits = cfg.getInt("credits");
 		int kills = cfg.getInt("kills");
 		int deaths = cfg.getInt("deaths");
 		String kit = cfg.getString("kit");
