+++ src/me/theremixpvp/ckitpvp/listeners/kits/SniperL.java
@@ -56,7 +56,7 @@ public class SniperL implements Listener {
 					}
 					Random rand = new Random();
 					int ri = rand.nextInt(10) + 1;
-					double rd = ri;
+					int rd = ri;
 					pd.setCredits(pd.credits() + rd);
 					pd.setKills(pd.kills() + 1);
 					PDUtils.getByName(vic.getName()).setDeaths(PDUtils.getByName(vic.getName()).deaths() + 1);
@@ -64,7 +64,7 @@ public class SniperL implements Listener {
 				} else {
 					Random rand = new Random();
 					int ri = rand.nextInt(10) + 1;
-					double rd = ri;
+					int rd = ri;
 					pd.setCredits(pd.credits() + rd);
 					pd.setKills(pd.kills() + 1);
 					p.sendMessage(ChatColor.GREEN + "You earned " + rd + " credits for killing " + e.getEntityType() + "!");
